PIP-1297

See merge request pipes/pipes!1230

+++ saas/console-api/src/tenants/TenantService.ts
@@ -7,11 +7,11 @@ import { authApp, usersService } from '../index';
 import Tenant = auth.Tenant;
 import { CollectionEnum } from '../enums/CollectionEnum';
 import UserCreationError from '../errors/UserCreationError';
-import BillingMongo from '../storage/mongo/Mongo';
+import Mongo from '../storage/mongo/Mongo';
 
 export default class TenantService {
 
-    public constructor(private readonly db: BillingMongo) {
+    public constructor(private readonly db: Mongo) {
     }
 
     public async getTenantList(): Promise<{ rows: unknown }> {
@@ -20,11 +20,13 @@ export default class TenantService {
         const rows = [] as ITenant[];
 
         await dbTenants.forEach((dbTenant) => {
-            const foundedTenant = tenants.tenants.find((tenant) => tenant.tenantId === dbTenant.tenantId);
+            const foundedTenant = tenants.tenants.find((tenant) => tenant.tenantId === dbTenant.gTenantId);
             if (foundedTenant) {
                 rows.push({
-                    instanceId: dbTenant.instanceId as string,
-                    ...foundedTenant,
+                    instances: dbTenant.instances,
+                    tenantId: dbTenant.tenantId,
+                    gTenantId: dbTenant.gTenantId,
+                    gTenant: foundedTenant,
                 } as ITenant);
             }
         });
@@ -32,20 +34,23 @@ export default class TenantService {
         return { rows: rows.map(this.mapTenantRecordToExport.bind(this)) };
     }
 
-    public async getTenant(tenantId: string): Promise<{ tenant: unknown }> {
+    public async getTenant(gTenantId: string): Promise<{ tenant: unknown }> {
         let tenant;
-        const dbTenant = await this.findOneTenant(tenantId);
+        const dbTenant = await this.findOneTenant(gTenantId);
 
         try {
+            const gTenant = await authApp.auth().tenantManager().getTenant(dbTenant.gTenantId);
             tenant = {
-                instanceId: dbTenant.instanceId,
-                ...await authApp.auth().tenantManager().getTenant(dbTenant.tenantId),
+                instances: dbTenant.instances,
+                tenantId: dbTenant.tenantId,
+                gTenantId: gTenant.tenantId,
+                gTenant,
             } as ITenant;
+
+            return { tenant: this.mapTenantRecordToExport(tenant) };
         } catch (e) {
             throw new TenantSearchError((e as Error).message);
         }
-
-        return { tenant: this.mapTenantRecordToExport(tenant) };
     }
 
     public async createTenant(createTenantRequest: ITenantCreateRequest): Promise<{ tenant: unknown }> {
@@ -62,32 +67,36 @@ export default class TenantService {
 
         const instanceId = crypto.randomBytes(5).toString('hex');
 
-        let tenant;
         try {
-            tenant = await this.updateCreatedGeneratedTenant(
+            const tenant = await this.updateCreatedGeneratedTenant(
                 generatedTenant,
                 createTenantRequest,
                 createUser,
-                instanceId,
+                [{ instanceId }],
+                `t-${instanceId}`,
             );
+
+            return { tenant: this.mapTenantRecordToExport(tenant) };
         } catch (e) {
             throw new TenantSearchError((e as Error).message);
         }
-
-        return { tenant: this.mapTenantRecordToExport(tenant) };
     }
 
     public async updateTenant(
-        tenantId: string,
+        gTenantId: string,
         updateTenantRequest: ITenantCreateRequest,
     ): Promise<{ tenant: unknown }> {
         let tenant;
-        const dbTenant = await this.findOneTenant(tenantId);
+        let gTenant;
+        const dbTenant = await this.findOneTenant(gTenantId);
 
         try {
+            gTenant = await authApp.auth().tenantManager().updateTenant(dbTenant.tenantId, updateTenantRequest);
             tenant = {
-                instanceId: dbTenant.instanceId,
-                ...await authApp.auth().tenantManager().updateTenant(dbTenant.tenantId, updateTenantRequest),
+                instances: dbTenant.instances,
+                tenantId: dbTenant.tenantId,
+                gTenantId: gTenant.tenantId,
+                gTenant,
             } as ITenant;
         } catch (e) {
             throw new TenantSearchError((e as Error).message);
@@ -96,14 +105,14 @@ export default class TenantService {
         return { tenant: this.mapTenantRecordToExport(tenant) };
     }
 
-    public async deleteTenant(tenantId: string): Promise<{ msg: string }> {
-        const dbTenant = await this.findOneTenant(tenantId);
-        const users = await authApp.auth().tenantManager().authForTenant(dbTenant.tenantId).listUsers();
+    public async deleteTenant(gTenantId: string): Promise<{ msg: string }> {
+        const dbTenant = await this.findOneTenant(gTenantId);
+        const users = await authApp.auth().tenantManager().authForTenant(dbTenant.gTenantId).listUsers();
         const userIds = users.users.map((user) => user.uid);
-        await authApp.auth().tenantManager().authForTenant(dbTenant.tenantId).deleteUsers(userIds);
+        await authApp.auth().tenantManager().authForTenant(dbTenant.gTenantId).deleteUsers(userIds);
 
         try {
-            await this.getTenantCollection().findOneAndDelete({ tenantId });
+            await this.getTenantCollection().findOneAndDelete({ tenantId: dbTenant.tenantId });
             await authApp.auth().tenantManager().deleteTenant(dbTenant.tenantId);
         } catch (e) {
             throw new TenantSearchError((e as Error).message);
@@ -114,20 +123,21 @@ export default class TenantService {
 
     private mapTenantRecordToExport(tenant: ITenant): unknown {
         return {
-            instanceId: tenant.instanceId,
+            instances: tenant.instances,
             tenantId: tenant.tenantId,
-            displayName: tenant.displayName ?? undefined,
-            emailSignInConfig: tenant.emailSignInConfig,
-            anonymousSignInEnabled: tenant.anonymousSignInEnabled,
+            gTenantId: tenant.gTenant?.tenantId,
+            displayName: tenant.gTenant?.displayName ?? undefined,
+            emailSignInConfig: tenant.gTenant?.emailSignInConfig,
+            anonymousSignInEnabled: tenant.gTenant?.anonymousSignInEnabled,
         };
     }
 
     private getTenantCollection(): Collection {
-        return this.db.getCollection(CollectionEnum.TENANT);
+        return this.db.getCloudCollection(CollectionEnum.TENANT);
     }
 
-    private async findOneTenant(tenantId: string): Promise<IDbTenant> {
-        return (await this.getTenantCollection().findOne({ tenantId })) as unknown as IDbTenant;
+    private async findOneTenant(gTenantId: string): Promise<ITenant> {
+        return (await this.getTenantCollection().findOne({ gTenantId })) as unknown as ITenant;
     }
 
     private async createGeneratedTenant(): Promise<Tenant> {
@@ -147,35 +157,41 @@ export default class TenantService {
         generatedTenant: Tenant,
         createTenantRequest: ITenantCreateRequest,
         createUser: boolean,
-        instanceId: string,
+        instances: IInstance[],
+        tenantId: string,
     ): Promise<ITenant> {
-        const tenant = await authApp.auth().tenantManager()
-            .updateTenant(generatedTenant.tenantId, { displayName: createTenantRequest.displayName }) as ITenant;
+        const gTenant = await authApp.auth().tenantManager()
+            .updateTenant(generatedTenant.tenantId, { displayName: createTenantRequest.displayName });
 
         if (createUser) {
             await usersService.createUser(
                 {
-                    tenantId: tenant.tenantId,
+                    tenantId: generatedTenant.tenantId,
                 },
                 {
                     email: createTenantRequest.email,
                     displayName: createTenantRequest.userDisplayName,
                 },
-                tenant.tenantId,
+                generatedTenant.tenantId,
             );
         }
 
-        await this.getTenantCollection().insertOne({ instanceId, tenantId: tenant.tenantId });
-        tenant.instanceId = instanceId;
+        const tenant = { instances, tenantId, gTenantId: gTenant.tenantId } as ITenant;
+        await this.getTenantCollection().insertOne(tenant);
+        tenant.gTenant = gTenant;
 
         return tenant;
     }
 
 }
 
-export interface ITenant extends Tenant, IDbTenant {}
+export interface ITenant {
+    instances: IInstance[];
+    tenantId: string;
+    gTenantId: string;
+    gTenant?: Tenant;
+}
 
-export interface IDbTenant {
+export interface IInstance {
     instanceId: string;
-    readonly tenantId: string;
 }
